+++ molmag_ac_gui/__main__.py
@@ -1,5 +1,10 @@
+"""
+Molmag AC GUI
+"""
+
 #std packages
 import ctypes
+import platform
 import sys
 
 #third-party packages
@@ -8,11 +13,13 @@ from PyQt5.QtWidgets import QApplication
 #local imports
 from . import ac_gui
 from . import process_ac
+from .__init__ import __version__
 
 if __name__ == '__main__':
     
-    myappid = 'AC Processing v1.0'
-    ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(myappid)
+    if platform.system() == 'Windows':
+        myappid = 'Molmag AC Gui {}'.format(__version__)
+        ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(myappid)
     
     app = QApplication(sys.argv)
     w = ac_gui.ACGui()