+++ HISTORY.md
@@ -1,3 +1,11 @@
+# v. 0.2.9
+Published on 26/10/2021
+Makes Windows-specific code only run on Windows
+
+# v. 0.2.8
+Published on 26/10/2021
+Updated read_options so SQUID-files can be read
+
 # v. 0.2.7
 Published on 25/10/2021